+++ packages/zowe-explorer/__tests__/__unit__/utils/AuthUtils.unit.test.ts
@@ -9,35 +9,50 @@
  *
  */
 
-import { ErrorCorrelator, Gui, imperative, ZoweExplorerApiType } from "@zowe/zowe-explorer-api";
+import { AuthHandler, ErrorCorrelator, Gui, imperative, ZoweExplorerApiType } from "@zowe/zowe-explorer-api";
 import { AuthUtils } from "../../../src/utils/AuthUtils";
 import { Constants } from "../../../src/configuration/Constants";
 import { MockedProperty } from "../../__mocks__/mockUtils";
 
 describe("AuthUtils", () => {
-    describe("promptForAuthError", () => {
+    describe("handleProfileAuthOnError", () => {
         it("should prompt for authentication", async () => {
-            const errorDetails = new imperative.ImperativeError({
+            const imperativeError = new imperative.ImperativeError({
                 errorCode: 401 as unknown as string,
                 msg: "All configured authentication methods failed",
             });
             const profile = { name: "aProfile", type: "zosmf" } as any;
+            const profilesCacheMock = new MockedProperty(Constants, "PROFILES_CACHE", {
+                value: {
+                    ssoLogin: jest.fn().mockImplementation(),
+                    promptCredentials: jest.fn().mockImplementation(),
+                } as any,
+                configurable: true,
+            });
             const correlateErrorMock = jest.spyOn(ErrorCorrelator.getInstance(), "correlateError");
-            const correlatedError = ErrorCorrelator.getInstance().correlateError(ZoweExplorerApiType.All, errorDetails, {
+            const errorCorrelation = ErrorCorrelator.getInstance().correlateError(ZoweExplorerApiType.All, imperativeError, {
                 templateArgs: {
                     profileName: profile.name,
                 },
             });
-            const promptForAuthenticationMock = jest
-                .spyOn(AuthUtils, "promptForAuthentication")
-                .mockImplementation(async () => Promise.resolve(true));
-            AuthUtils.promptForAuthError(errorDetails, profile);
-            expect(correlateErrorMock).toHaveBeenCalledWith(ZoweExplorerApiType.All, errorDetails, {
+            const isUsingTokenAuthMock = jest.spyOn(AuthUtils, "isUsingTokenAuth").mockResolvedValueOnce(false);
+            const promptForAuthenticationMock = jest.spyOn(AuthHandler, "promptForAuthentication").mockResolvedValueOnce(true);
+            await AuthUtils.handleProfileAuthOnError(imperativeError, profile);
+            expect(correlateErrorMock).toHaveBeenCalledWith(ZoweExplorerApiType.All, imperativeError, {
                 templateArgs: {
                     profileName: profile.name,
                 },
             });
-            expect(promptForAuthenticationMock).toHaveBeenCalledWith(errorDetails, profile, correlatedError);
+            expect(promptForAuthenticationMock).toHaveBeenCalledWith(
+                profile,
+                expect.objectContaining({
+                    imperativeError,
+                    errorCorrelation,
+                    isUsingTokenAuth: false,
+                })
+            );
+            profilesCacheMock[Symbol.dispose]();
+            isUsingTokenAuthMock.mockRestore();
         });
     });
     describe("promptForSsoLogin", () => {