+++ packages/zowe-explorer/src/configuration/Profiles.ts
@@ -26,6 +26,7 @@ import {
     FileManagement,
     IRegisterClient,
     Types,
+    AuthHandler,
 } from "@zowe/zowe-explorer-api";
 import { SettingsConfig } from "./SettingsConfig";
 import { Constants } from "./Constants";
@@ -828,6 +829,7 @@ export class Profiles extends ProfilesCache {
                         comment: ["Service profile name"],
                     })
                 );
+                AuthHandler.unlockProfile(serviceProfile, true);
             }
             return loginOk;
         } catch (err) {